Added percent change when a list item is clicked

+++ states.go
@@ -2,5 +2,5 @@ package main
 
 type State struct {
 	FullName      string //includes state name and pop
-	PercentChange float32
+	PercentChange float64
 }
